refactor: Move domain storage logic to data modules
