--- core/localstorage/src/main/java/com/hilingual/core/localstorage/UserInfoManager.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.core.localstorage
-
-interface UserInfoManager {
-    suspend fun saveRegisterStatus(isCompleted: Boolean)
-    suspend fun getRegisterStatus(): Boolean
-    suspend fun saveOtpVerified(isVerified: Boolean)
-    suspend fun isOtpVerified(): Boolean
-    suspend fun clear()
-}
