+++ data/onboarding/src/main/java/com/hilingual/data/onboarding/localstorage/OnboardingLocalDataSourceImpl.kt
@@ -1,19 +1,4 @@
-/*
- * Copyright 2026 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.core.localstorage
+package com.hilingual.data.onboarding.localstorage
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
@@ -22,9 +7,9 @@ import androidx.datastore.preferences.core.edit
 import javax.inject.Inject
 import kotlinx.coroutines.flow.first
 
-class OnboardingStateManagerImpl @Inject constructor(
+class OnboardingLocalDataSourceImpl @Inject constructor(
     private val dataStore: DataStore<Preferences>
-) : OnboardingStateManager {
+) : OnboardingLocalDataSource {
     private object PreferencesKeys {
         val IS_HOME_ONBOARDING_COMPLETED = booleanPreferencesKey("is_home_onboarding_completed")
         val IS_SPLASH_ONBOARDING_COMPLETED = booleanPreferencesKey("is_splash_onboarding_completed")
