+++ data/user/src/main/java/com/hilingual/data/user/localstorage/UserLocalDataSource.kt
@@ -1,0 +1,7 @@
+package com.hilingual.data.user.localstorage
+
+interface UserLocalDataSource {
+    suspend fun saveRegisterStatus(isCompleted: Boolean)
+    suspend fun getRegisterStatus(): Boolean
+    suspend fun clear()
+}
